Chore: 시간 관련 타입 한 파일에 정리
