--- iOS/Targets/Cherrybnb/Sources/Common/Second.swift
@@ -1,11 +1,0 @@
-//
-//  Second.swift
-//  Cherrybnb
-//
-//  Created by Bumgeun Song on 2022/05/25.
-//  Copyright © 2022 Codesquad. All rights reserved.
-//
-
-import Foundation
-
-typealias Second = Float
